Making number of pixels a size_t rather than int. Extra debug statement added as well.
git-svn-id: https://svn.atnf.csiro.au/askapsdp/Src/trunk@2471 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Components/Analysis/simulations/trunk/FITS/FITSfile.h
@@ -162,7 +162,7 @@ namespace askap {
                     /// @brief The axis dimensions
                     std::vector<int> itsAxes;
                     /// @brief The number of pixels in the image
-                    int itsNumPix;
+                    size_t itsNumPix;
                     /// @brief The section of the image in which to place sources - defaults to the null section of the appropriate dimensionality, and needs to be set explicitly via setSection()
                     duchamp::Section itsSourceSection;
